[Refactor] HomePage 스타일 분리
