+++ src/Main.jsx
@@ -1,12 +1,12 @@
 import { Route, BrowserRouter, Routes } from "react-router-dom";
 import { createGlobalStyle } from "styled-components";
 //
-import LandingPage from "../pages/LandingPage/LandingPage.jsx";
-import App from "../App";
-import HomePage from "../pages/HomePage";
-import AddItem from "../pages/AddItem/AddItem.jsx";
+import LandingPage from "./pages/LandingPage/LandingPage.jsx";
+import App from "./App.js";
+import HomePage from "./pages/HomePage";
+import AddItem from "./pages/AddItem/AddItem.jsx";
 //
-import useWindowSize from "../hooks/useWindowSize";
+import useWindowSize from "./hooks/useWindowSize";
 //
 const GlobalStyle = createGlobalStyle`
   * {
