+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Main from "./components/Main";
+import Main from "./Main.jsx";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(<Main />);
